movie crud section

+++ src/utils/delete-file.ts
@@ -1,9 +1,7 @@
 import * as fs from 'fs';
 import { join } from 'path';
 
-export function deleteFileInPublic(
-  fileAddress: string,
-) {
+export function deleteFileInPublic(fileAddress: string) {
   if (fileAddress) {
     const filePath = join(
       __dirname,
@@ -13,7 +11,20 @@ export function deleteFileInPublic(
       'musics',
       fileAddress,
     );
-    if (fs.existsSync(filePath))
-      fs.unlinkSync(filePath);
+    if (fs.existsSync(filePath)) fs.unlinkSync(filePath);
+  }
+}
+
+export function deleteImageInPublic(imageAddress: string) {
+  if (imageAddress) {
+    const filePath = join(
+      __dirname,
+      '..',
+      '..',
+      'uploads',
+      'films',
+      imageAddress,
+    );
+    if (fs.existsSync(filePath)) fs.unlinkSync(filePath);
   }
 }
